+++ CHANGELOG.md
@@ -5,7 +5,9 @@
   - Use more optimal SQL when polling workflows when database supports update returning syntax
   - Only rollback poll operation when no workflows could be allocated for executing (when multiple pollers compete for same workflows)
   - Allow configuring executor queue length with _nflow.dispatcher.executor.queue.size_
+  - Added service to stop the execution of the workflow instance
 - nflow-rest:
+  - Added service to stop the execution of the workflow instance
   - Add support for user-provided action description when updating a workflow instance
 
 ## 1.2.0 (2014-12-23)